requirements.txt
pandas>=1.5.0
numpy>=1.20.0
dhanhq>=1.3.0
requests>=2.28.0
python-dateutil>=2.8.0
flask>=2.3.0
flask-cors>=4.0.0
openpyxl>=3.1.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
schedule>=1.2.0